refactor(migration): one implementation of the plumbing, anchored to the script
`migrationFixture/config.ts` carried a second implementation of every
plumbing concern the CLI has: chain synthesis, deployment loading, client
construction. The two had drifted, and the drift was a bug — the fixture's
v1 deployment roots resolved against the process working directory while
the CLI's resolve against the script, so `fixture` commands failed to find
artifacts that every other command found, depending only on where the
operator stood when they ran it. It was the sole cwd-relative path in the
whole script set.

Networks, deployment roots, chains, clients, artifact loaders and the
numeric parser now have one home, which both the CLI and the fixture
import. The fixture's `optionalV1Deployment` also stops swallowing parse
errors: only an absent artifact is an answer, and a corrupt one now says
so rather than reading as "not deployed".

The module reaches back into nothing, so the fixture subtree can import it
without closing a cycle — which importing from `migration.ts` would have
done, since `migration.ts` already imports the fixture.
